% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.R
\name{close}
\alias{close}
\alias{close.tiledb_array}
\alias{close.TileDBArray}
\alias{close.tiledb_group}
\alias{close.TileDBGroup}
\title{Close a TileDB Resource}
\usage{
\method{close}{tiledb_array}(con, ...)

\method{close}{TileDBArray}(con, ...)

\method{close}{tiledb_group}(con, ...)

\method{close}{TileDBGroup}(con, ...)
}
\arguments{
\item{con}{An \code{R} object that contains a \code{TileDB} resource pointer.}

\item{...}{Other arguments passed to methods. Not used.}
}
\value{
Invisibly, a logical \code{TRUE} on success.
}
\description{
Close a \code{\link[tiledb:tiledb_array]{tiledb::tiledb_array()}}, \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group()}}, \link{TileDBArray}
or \link{TileDBGroup} object.
}
