% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{delete_metadata}
\alias{delete_metadata}
\alias{delete_metadata.TileDBArray}
\alias{delete_metadata.TileDBGroup}
\alias{delete_metadata.tiledb_array}
\alias{delete_metadata.tiledb_group}
\alias{delete_metadata.character}
\title{Delete TileDB Metadata}
\usage{
\method{delete_metadata}{TileDBArray}(x, keys)

\method{delete_metadata}{TileDBGroup}(x, keys)

\method{delete_metadata}{tiledb_array}(x, keys)

\method{delete_metadata}{tiledb_group}(x, keys)

\method{delete_metadata}{character}(x, keys)
}
\arguments{
\item{x}{An \code{R} object that points to a \code{TileDB} resource whose
metadata are to be accessed.}

\item{keys}{A character vector of metadata key names to be accessed.}
}
\value{
A logical \code{TRUE}, invisibly.
}
\description{
Delete metadata using keys names for a \code{TileDB} array or group.
}
\details{
\code{delete_metadata()} works similar to \code{\link[=metadata<-]{metadata<-()}} with \code{NULL} value
but works with a character vector of keys.

The character method requires a valid URI path.

The methods will not alter the mode of the \code{TileDB} object; also, the object
will be opened temporarily to access the metadata if it is closed.
}
\seealso{
\code{\link[=set_metadata]{set_metadata()}} and \code{\link[=metadata]{metadata()}}.
}
