% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demo_array_UCBAdmissions}
\alias{demo_array_UCBAdmissions}
\title{Create a demo array for \code{UCBAdmissions} dataset}
\usage{
demo_array_UCBAdmissions(uri)
}
\arguments{
\item{uri}{The uri path.}
}
\description{
Create a demo array for \code{UCBAdmissions} dataset
}
