% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demo_UCBAdmissions_array}
\alias{demo_UCBAdmissions_array}
\title{Create array for \code{UCBAdmissions} dataset}
\usage{
demo_UCBAdmissions_array(uri)
}
\arguments{
\item{uri}{The uri path.}
}
\description{
Create array for \code{UCBAdmissions} dataset
}
