% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_timestamps.R
\name{group_timestamps}
\alias{group_timestamps}
\alias{group_timestamps.tiledb_ctx}
\alias{group_timestamps.tiledb_group}
\alias{group_timestamps.tiledb_config}
\alias{group_timestamps.TileDBGroup}
\title{Get Group Timestamps}
\usage{
\method{group_timestamps}{tiledb_ctx}(object, tz = "", ...)

\method{group_timestamps}{tiledb_group}(object, tz = "", ...)

\method{group_timestamps}{tiledb_config}(object, tz = "", ...)

\method{group_timestamps}{TileDBGroup}(object, from = c("ctx", "cfg"),
  tz = "", ...)
}
\arguments{
\item{object}{An \code{R} object that contains a \code{TileDB} resource pointer.}

\item{tz}{A character string for the time zone specification to be used
for the conversion in print method only. Defaults to \code{\link[=Sys.timezone]{Sys.timezone()}}.}

\item{...}{Other arguments passed to methods. Not used.}

\item{from}{TileDBGroup's source of timestamps: either from context \code{ctx}, or
group object configuration \code{cfg}.}
}
\value{
An object of class \code{group_timestamps} that is a list with
the time range the group is opened at.
}
\description{
Get Group Timestamps
}
\seealso{
\code{\link[=array_timestamps]{array_timestamps()}}, \code{\link[=set_tiledb_timestamp]{set_tiledb_timestamp()}}
}
