% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{metadata}
\alias{metadata}
\alias{metadata<-}
\alias{metadata.TileDBArray}
\alias{metadata.TileDBGroup}
\alias{metadata.tiledb_array}
\alias{metadata.tiledb_group}
\alias{metadata.character}
\alias{metadata<-.TileDBArray}
\alias{metadata<-.TileDBGroup}
\alias{metadata<-.tiledb_array}
\alias{metadata<-.tiledb_group}
\alias{metadata<-.character}
\title{TileDB Object Metadata}
\usage{
metadata(x, which)

metadata(x, which) <- value

\method{metadata}{TileDBArray}(x, which)

\method{metadata}{TileDBGroup}(x, which)

\method{metadata}{tiledb_array}(x, which)

\method{metadata}{tiledb_group}(x, which)

\method{metadata}{character}(x, which)

\method{metadata}{TileDBArray}(x, which) <- value

\method{metadata}{TileDBGroup}(x, which) <- value

\method{metadata}{tiledb_array}(x, which) <- value

\method{metadata}{tiledb_group}(x, which) <- value

\method{metadata}{character}(x, which) <- value
}
\arguments{
\item{x}{An \code{R} object that points to a \code{TileDB} resource whose
metadata are to accessed.}

\item{which}{A non-empty character string specifying which metadata key
is to be accessed.}

\item{value}{An object, the new value of the metadata, or \code{NULL} to remove
the key. Note that character vectors should be of length one (scalar).}
}
\value{
For the extractor, the key value of the metadata matched, or \code{NULL}
if no exact match is found.
}
\description{
Get or set a metadata key for a \code{TileDB} array or group.
}
\details{
These functions work similar to \code{\link[=attr]{attr()}} and provide access to a
single metadata key on an TileDB object. The replacement form (setter)
puts a metadata key with the value specified or creates a new metadata
with the value given.

The character method is intended for a valid URI path.

The methods will not alter the mode of the \code{TileDB} object.

For the extractor the object will be opened temporarily to access the
metadata if it is closed.

For the replacement, the object will be opened in write mode temporarily
and on exit will revert back to previous mode; in the case where the object
is already in write mode then it will be reopened at the same mode in order
to flush the metadata on disk.
}
