% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{new_context}
\alias{new_context}
\title{Create a new TileDB Context}
\usage{
new_context(cfg = NULL)
}
\arguments{
\item{cfg}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}}.}
}
\description{
Compared to \code{tiledb_ctx()}, this variant does not cache the
context and does not accept a character vector of configuration parameters.
}
\keyword{internal}
