% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.tiledb-package.R
\docType{package}
\name{R6.tiledb-package}
\alias{R6.tiledb}
\alias{R6.tiledb-package}
\title{R6.tiledb: R6 Classes for TileDB Arrays and Groups}
\description{
Provides foundational R6 classes to interface with 'TileDB' storage engine using 'tiledb' package. The goal is to offer a set of minimum classes based on the R6 object oriented system in order to make it easier to work with 'TileDB' Arrays and Groups. The base classes 'TileDBArray' and 'TileDBGroup' can be easily extended to new subclasses to represent and support any data model.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cgiachalis/R6.tiledb}
  \item \url{https://cgiachalis.github.io/R6.tiledb/}
  \item Report bugs at \url{https://github.com/cgiachalis/R6.tiledb/issues}
}

}
\author{
\strong{Maintainer}: Constantinos Giachalis \email{26255858+cgiachalis@users.noreply.github.com}

Authors:
\itemize{
  \item Constantinos Giachalis \email{26255858+cgiachalis@users.noreply.github.com}
}

Other contributors:
\itemize{
  \item TileDB, Inc (Initial code and tests for R6 classes from 'TileDB-SOMA/apis/r' GitHub repo.) [contributor]
}

}
\keyword{internal}
