% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_config_params.R
\name{set_config_params}
\alias{set_config_params}
\alias{unset_config_params}
\title{Modify Config Parameters}
\usage{
set_config_params(cfg, keyval)

unset_config_params(cfg, keys)
}
\arguments{
\item{cfg}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}}.}

\item{keyval}{A named character vector with configuration parameters.}

\item{keys}{A character with configuration keys to unset.}
}
\value{
The modified \code{tiledb_config} object.
}
\description{
These functions will set or unset the parameters of TileDB Config object
and return the modified configuration object.
}
\examples{
\dontrun{
origparams <- c("sm.consolidation.timestamp_start" = "100",
                "sm.consolidation.timestamp_end" = "20000")

# Create a context with a config setting the params above
cfg <- tiledb::tiledb_config(origparams)
ctx <- tiledb::tiledb_ctx(config = cfg)

# Do something with ctx

# Get configs from context
cfg <- tiledb::config(ctx)

# New params
params <-  c("sm.consolidation.timestamp_start" = "100",
             "sm.consolidation.timestamp_end" = "10000")

cfg <- set_config_params(cfg, params)

cfg["sm.consolidation.timestamp_end"] # "10000"

# Update context with modified config object
ctx <- tiledb::tiledb_ctx(config = cfg)


# Reset consolidation timestamps
cfg <- unset_config_params(cfg, names(params))

# check the default values
cfg["sm.consolidation.timestamp_start"] # "0"
cfg["sm.consolidation.timestamp_end"]   # "18446744073709551615"

# Update context with modified config object
ctx <- tiledb::tiledb_ctx(config = cfg)

}
}
\seealso{
\code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}}
}
