% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_consolidation_tstamps.R
\name{set_consolidation_tstamps}
\alias{set_consolidation_tstamps}
\alias{unset_consolidation_tstamps}
\title{Modify Consolidation Start/End Timestamps}
\usage{
set_consolidation_tstamps(cfg, start_time = NULL, end_time = NULL)

unset_consolidation_tstamps(cfg)
}
\arguments{
\item{cfg}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}}.}

\item{start_time, end_time}{Optional time stamp values. A date time objects
of class \code{POSIXt}.}
}
\value{
The modified \code{tiledb_config} object.
}
\description{
Set or unset consolidation timestamps to the given configuration object.
}
\examples{
\dontrun{
cfg <- tiledb::tiledb_config()
cfg <- set_consolidation_tstamps(cfg, as.POSIXct(1), as.POSIXct(1000))
cfg["sm.consolidation.timestamp_start"] # 1000 (ms)

# reset
cfg <- unset_consolidation_tstamps(cfg)
cfg["sm.consolidation.timestamp_start"] # 0
}
}
\seealso{
\code{\link[=set_config_params]{set_config_params()}}
}
