% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_group_timestamps.R
\name{set_group_timestamps}
\alias{set_group_timestamps}
\alias{unset_group_timestamps}
\title{Modify Group Start/End Open Timestamps}
\usage{
set_group_timestamps(cfg, start_time = NULL, end_time = NULL)

unset_group_timestamps(cfg)
}
\arguments{
\item{cfg}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}}.}

\item{start_time, end_time}{Optional time stamp values. A date time objects
of class \code{POSIXt}.}
}
\value{
The modified \code{tiledb_config} object.
}
\description{
Set or unset group opening start, end timestamps to the given
configuration object.
}
\examples{
cfg <- tiledb::tiledb_config()
cfg <- set_group_timestamps(cfg, as.POSIXct(1), as.POSIXct(1000))
cfg["sm.group.timestamp_start"] # 1000 (ms)

# reset
cfg <- unset_group_timestamps(cfg)
cfg["sm.group.timestamp_start"] # 0

}
\seealso{
\code{\link[=set_config_params]{set_config_params()}}
}
