% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{set_metadata}
\alias{set_metadata}
\alias{set_metadata.TileDBArray}
\alias{set_metadata.TileDBGroup}
\alias{set_metadata.tiledb_array}
\alias{set_metadata.tiledb_group}
\alias{set_metadata.character}
\title{Set TileDB Metadata}
\usage{
\method{set_metadata}{TileDBArray}(x, keys, timestamp = NULL)

\method{set_metadata}{TileDBGroup}(x, keys, timestamp = NULL)

\method{set_metadata}{tiledb_array}(x, keys, timestamp = NULL)

\method{set_metadata}{tiledb_group}(x, keys, timestamp = NULL)

\method{set_metadata}{character}(x, keys, timestamp = NULL)
}
\arguments{
\item{x}{An \code{R} object that points to a \code{TileDB} resource whose
metadata are to written.}

\item{keys}{A named list of key value metadata.}

\item{timestamp}{Optional datetime object of class \code{"POSIXct"} to write
at this timestamp.}
}
\value{
A named list of class \code{tdb_metadata}.
}
\description{
Set metadata as a list of key-value pairs for a \code{TileDB} array or group.
Optionally, you can write metadata at specific point in time
(time-travelling).
}
\details{
This functions work similar to \code{\link[=metadata<-]{metadata<-()}} but works with
a list of key value pairs and always returns a list.

The optional argument \code{timestamp} can be used to set metadata at specific
point in time.

The character method is intended for a valid URI path.

The methods will not alter the mode of the \code{TileDB} object.

The object will be opened in write mode temporarily and on exit will
revert back to previous mode; in the case where the object is already
in write mode then it will be reopened at the same mode in order
to flush the metadata on disk.
}
\seealso{
\code{\link[=fetch_metadata]{fetch_metadata()}} and \code{\link[=metadata]{metadata()}}.
}
