% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tiledb_timestamp.R
\name{set_tiledb_timestamp}
\alias{set_tiledb_timestamp}
\title{Set a \code{TileDB} time-stamp}
\usage{
set_tiledb_timestamp(ts_start, ts_end, tz = "UTC")
}
\arguments{
\item{ts_start, ts_end}{An object coercible to \code{POSIXct} using \code{\link[=as.POSIXct]{as.POSIXct()}}.}

\item{tz}{A character string for the time zone specification to be used
for the conversion.}
}
\value{
An list object of class \code{tiledb_timestamp}.
}
\description{
Define a list of start,end timestamps intended for opening a \code{TileDB}
object, either at initialisation or via active field \code{tiledb_timestamp}.
}
\examples{
# Character input
set_tiledb_timestamp("2025-01-01", "2025-08-01")

# Numeric input
set_tiledb_timestamp(1000000, 1000000*4)

}
\seealso{
\link{TileDBObject}
}
