% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdb_array.R
\name{tdb_array_create}
\alias{tdb_array_create}
\title{Create a TileDB Array}
\usage{
tdb_array_create(uri, sch, mode = "WRITE", ctx = NULL)
}
\arguments{
\item{uri}{URI path for the \code{TileDB} object.}

\item{sch}{A TileDB schema. See constructor \code{\link[tiledb:tiledb_array_schema]{tiledb::tiledb_array_schema()}}.}

\item{mode}{Mode to open: either \code{"READ"} or \code{"WRITE"} (default).}

\item{ctx}{A TileDB Context. See \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}}.}
}
\value{
A \code{TileDBArrayExp}, \code{R6} object.
}
\description{
Given a TileDB schema, it will create an array and instantiate a
\code{TileDBArrayExp} object. The array will be opened at the given mode
and kept opened. It can be accessed via active field \verb{$object}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
suppressMessages(library(tiledb))

# construct schema
schema <- tiledb_array_schema(
 dom = tiledb_domain(dims = tiledb_dim("id", c(1L, 4L), 4L, "INT32")),
 attrs = tiledb_attr("model", type = "ASCII"))

# uri path
uri <- tempfile()

# create and open array at WRITE mode
arrobj <- tdb_array_create(uri, sch = schema, mode = "WRITE")

arrobj$schema_info()

arrobj$mode # WRITE
\dontshow{\}) # examplesIf}
}
