% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdb_group.R
\name{tdb_group}
\alias{tdb_group}
\title{Open \code{TileDBGroupExp}}
\usage{
tdb_group(uri, mode = "READ", ctx = NULL, tiledb_timestamp = NULL)
}
\arguments{
\item{uri}{URI path for the \code{TileDB} object.}

\item{mode}{Mode to open : either \verb{"READ" (default) or "WRITE"}.}

\item{ctx}{Optional \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}} object.}

\item{tiledb_timestamp}{Set a \code{TileDB} timestamp range that
the resource will be opened at. Effective only for \code{"READ"} mode.
Valid options:
\itemize{
\item A \code{NULL} value (default)
\item An \code{R} object coercible to \code{POSIXct} with length 1 which used for end timestamp,
or length 2 with start, end timestamps
\item An object of class \code{tiledb_timestamp}. See \code{\link[=set_tiledb_timestamp]{set_tiledb_timestamp()}}
}

Also, it can be set through active field \verb{$tiledb_timestamp}.}
}
\value{
A \code{TileDBGroupExp}, \code{R6} object.
}
\description{
Functional interface that initialises a \link{TileDBGroupExp} instance and opens
a group object at \code{READ} or \code{WRITE} mode.
}
\section{Active bindings}{

\itemize{
\item \code{ctx} : A TileDB Context. See \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}}
\item \code{tiledb_timestamp} : A \code{TileDB} timestamp range that
the resource will be opened at. See \code{\link[=set_tiledb_timestamp]{set_tiledb_timestamp()}}
\item \code{uri} : The URI of the \code{TileDB} object
\item \code{mode}: Get the mode of the object: one of the following:
\code{"CLOSED"}, \code{"READ"} or \code{"WRITE"}
\item \code{object_type} : The \code{TileDB} object type: \code{"ARRAY"},\code{"GROUP"} or \code{"INVALID"}
\item \code{object} : Access the underlying \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group()}} object
\item \code{members}: Access the list of group members.
}
}

\section{Methods}{


\strong{Public Methods}

\if{html}{\out{
<ul>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="new"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-new'><code>$new()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="class"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-class'><code>$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="is_open"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-is_open'><code>$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="exists"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-exists'><code>$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="get_metadata"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-get_metadata'><code>$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="set_metadata"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-set_metadata'><code>$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="create"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-create'><code>$create()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="open"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-open'><code>$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="close"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-close'><code>$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="remove"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-remove'><code>$remove()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="delete"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-delete'><code>$delete()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="count_members"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-count_members'><code>$count_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="get_members_df"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-get_members_df'><code>$get_members_df()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="get_member"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-get_member'><code>$get_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="set_member"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-set_member'><code>$set_member()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="names"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-names'><code>$names()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="member_exists"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-member_exists'><code>$member_exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="print"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-print'><code>$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroup" data-id="dump"><a href='../../R6.tiledb/html/TileDBGroup.html#method-TileDBGroup-dump'><code>$dump()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroupExp" data-id="has_non_members"><a href='../../R6.tiledb/html/TileDBGroupExp.html#method-TileDBGroupExp-has_non_members'><code>$has_non_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroupExp" data-id="non_members"><a href='../../R6.tiledb/html/TileDBGroupExp.html#method-TileDBGroupExp-non_members'><code>$non_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroupExp" data-id="prune_non_members"><a href='../../R6.tiledb/html/TileDBGroupExp.html#method-TileDBGroupExp-prune_non_members'><code>$prune_non_members()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroupExp" data-id="delete_group"><a href='../../R6.tiledb/html/TileDBGroupExp.html#method-TileDBGroupExp-delete_group'><code>$delete_group()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroupExp" data-id="walk_group"><a href='../../R6.tiledb/html/TileDBGroupExp.html#method-TileDBGroupExp-walk_group'><code>$walk_group()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBGroupExp" data-id="dir_tree"><a href='../../R6.tiledb/html/TileDBGroupExp.html#method-TileDBGroupExp-dir_tree'><code>$dir_tree()</code></a></span></li>
</ul>
}}
}

\examples{
\dontrun{
 # uri path
 uri <- tempdir()

 obj <- TileDBGroup$new(uri)

 obj$create()

 obj$close()

 # new instance
 newobj <- tdb_group(uri)

 newobj$is_open() # TRUE

 newobj$mode # "READ"
}

}
