% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdb_group.R
\name{tdb_group_create}
\alias{tdb_group_create}
\title{Create a TileDB Group}
\usage{
tdb_group_create(uri, mode = "WRITE", ctx = NULL)
}
\arguments{
\item{uri}{URI path for the \code{TileDB} object.}

\item{mode}{Mode to open: either \code{"READ"} or \code{"WRITE"} (default).}

\item{ctx}{A TileDB Context. See \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}}.}
}
\value{
A \code{TileDBGroupExp}, \code{R6} object.
}
\description{
Create a group and instantiate a \code{TileDBGroupExp} object. The group will be
opened at the given mode and kept opened. It can be accessed via active field
\verb{$object}.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# uri path
uri <- tempfile()

# create and open array at WRITE mode
grpobj <- tdb_group_create(uri,mode = "WRITE")

grpobj$mode # WRITE

grpobj$count_members() # 0
\dontshow{\}) # examplesIf}
}
