% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArrayExtended.R
\name{TileDBArrayExtended}
\alias{TileDBArrayExtended}
\title{Generate a \code{TileDBArrayExtended} Object}
\value{
An object of class \code{TileDBArrayExtented}.
}
\description{
This class inherits from \link{TileDBArray} and offers additional methods to
operate on the array.

\strong{TODO} add section with methods: enum, consol/vac, fragment
\subsection{Initialization}{

A new \code{TileDBArrayExtended} instance is initialized using the \code{new()} method.
Alternatively use \code{\link[=tdb_array]{tdb_array()}} to create an instance and open the array at
\code{READ} mode.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ # uri path
 uri <- tempdir()
 # new instance
 obj <- TileDBArrayExtended$new(uri = uri)
 # does array exist at this uri
 obj$exists() # FALSE

 unlink(uri)
}\if{html}{\out{</div>}}
}
}
\section{Super classes}{
\code{\link[R6.tiledb:TileDBObject]{R6.tiledb::TileDBObject}} -> \code{\link[R6.tiledb:TileDBArray]{R6.tiledb::TileDBArray}} -> \code{TileDBArrayExtended}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fragments_object}}{Access the \link{TileDBFragments} instance for this
array.}

\item{\code{schema_version}}{Retrieve the schema version for this array.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBArrayExtended-new}{\code{TileDBArrayExtended$new()}}
\item \href{#method-TileDBArrayExtended-reopen}{\code{TileDBArrayExtended$reopen()}}
\item \href{#method-TileDBArrayExtended-enum_columns}{\code{TileDBArrayExtended$enum_columns()}}
\item \href{#method-TileDBArrayExtended-enum_levels}{\code{TileDBArrayExtended$enum_levels()}}
\item \href{#method-TileDBArrayExtended-has_enums}{\code{TileDBArrayExtended$has_enums()}}
\item \href{#method-TileDBArrayExtended-consolidate}{\code{TileDBArrayExtended$consolidate()}}
\item \href{#method-TileDBArrayExtended-consolidate_async}{\code{TileDBArrayExtended$consolidate_async()}}
\item \href{#method-TileDBArrayExtended-vacuum}{\code{TileDBArrayExtended$vacuum()}}
\item \href{#method-TileDBArrayExtended-vacuum_async}{\code{TileDBArrayExtended$vacuum_async()}}
\item \href{#method-TileDBArrayExtended-drop_attribute}{\code{TileDBArrayExtended$drop_attribute()}}
\item \href{#method-TileDBArrayExtended-frag_num}{\code{TileDBArrayExtended$frag_num()}}
\item \href{#method-TileDBArrayExtended-frag_to_vacuum}{\code{TileDBArrayExtended$frag_to_vacuum()}}
\item \href{#method-TileDBArrayExtended-frag_dump}{\code{TileDBArrayExtended$frag_dump()}}
\item \href{#method-TileDBArrayExtended-frag_uris}{\code{TileDBArrayExtended$frag_uris()}}
\item \href{#method-TileDBArrayExtended-schema_upgrade}{\code{TileDBArrayExtended$schema_upgrade()}}
\item \href{#method-TileDBArrayExtended-clone}{\code{TileDBArrayExtended$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="class"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-class'><code>R6.tiledb::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="exists"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-exists'><code>R6.tiledb::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="is_open"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-is_open'><code>R6.tiledb::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="mode"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-mode'><code>R6.tiledb::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="attributes"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-attributes'><code>R6.tiledb::TileDBArray$attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="attrnames"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-attrnames'><code>R6.tiledb::TileDBArray$attrnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="close"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-close'><code>R6.tiledb::TileDBArray$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="colnames"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-colnames'><code>R6.tiledb::TileDBArray$colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="dimensions"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-dimensions'><code>R6.tiledb::TileDBArray$dimensions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="dimnames"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-dimnames'><code>R6.tiledb::TileDBArray$dimnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="get_metadata"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-get_metadata'><code>R6.tiledb::TileDBArray$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="open"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-open'><code>R6.tiledb::TileDBArray$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="print"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-print'><code>R6.tiledb::TileDBArray$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="schema"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-schema'><code>R6.tiledb::TileDBArray$schema()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="schema_info"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-schema_info'><code>R6.tiledb::TileDBArray$schema_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="set_metadata"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-set_metadata'><code>R6.tiledb::TileDBArray$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="tiledb_array"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-tiledb_array'><code>R6.tiledb::TileDBArray$tiledb_array()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TileDBArrayExtended} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$new(uri, ctx = NULL, tiledb_timestamp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI path for the \code{TileDB} Array.}

\item{\code{ctx}}{Optional \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}} object.}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-reopen"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-reopen}{}}}
\subsection{Method \code{reopen()}}{
Close and reopen the TileDB object in a new mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$reopen(mode = c("READ", "WRITE"),
  tiledb_timestamp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{New mode to open the object in; choose from: \code{"READ"} or \code{"WRITE"}.}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-enum_columns"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-enum_columns}{}}}
\subsection{Method \code{enum_columns()}}{
Retrieve factor columns (attributes).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$enum_columns()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with factor columns (enum attributes).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-enum_levels"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-enum_levels}{}}}
\subsection{Method \code{enum_levels()}}{
Retrieve factor levels for a given attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$enum_levels(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An attribute name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector with levels (enum values).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-has_enums"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-has_enums}{}}}
\subsection{Method \code{has_enums()}}{
Checks array for factor columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$has_enums()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A boolean. \code{TRUE} indicating the array has factor columns and
\code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
Consolidates the fragments of the array into
a single fragment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$consolidate(cfg = NULL, start_time = NULL,
  end_time = NULL, mode = c("fragments", "commits", "fragment_meta",
  "array_meta"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}} to set parameters
for the consolidation. When \code{NULL} (default) the configuration parameters will
be retrieve from object context.}

\item{\code{start_time, end_time}}{Optional time stamp values. A date time objects
of class \code{POSIXlt}. If not provided, the default values from configuration
object will be used.}

\item{\code{mode}}{The consolidate mode, one of the following:
\itemize{
\item \code{"fragments"}: - consolidate all fragments (default)
\item \code{"commits"}: - consolidate all commit files
\item \code{"fragment_meta"}: - consolidate only fragment metadata footers to a single file
\item \code{"array_meta"}: - consolidate array metadata only
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE}, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-consolidate_async"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-consolidate_async}{}}}
\subsection{Method \code{consolidate_async()}}{
Consolidate fragments of the array into a single fragment
asynchronously.

The consolidation will run in a separate R process in a clean environment.

\strong{Note this function requires the \href{https://cran.r-project.org/web/packages/mirai/index.html}{mirai} package}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$consolidate_async(cfg = NULL, start_time = NULL,
  end_time = NULL, mode = c("fragments", "commits", "fragment_meta",
  "array_meta"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}} to set parameters
for the consolidation. When \code{NULL} (default) the configuration parameters will
be retrieve from object context.}

\item{\code{start_time, end_time}}{Optional time stamp values. A date time objects
of class \code{POSIXlt}. If not provided, the default values from configuration
object will be used.}

\item{\code{mode}}{The consolidate mode, one of the following:
\itemize{
\item \code{"fragments"}: - consolidate all fragments (default)
\item \code{"commits"}: - consolidate all commit files
\item \code{"fragment_meta"}: - consolidate only fragment metadata footers to a single file
\item \code{"array_meta"}: - consolidate array metadata only
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This function will return a \code{\link[mirai:mirai]{mirai::mirai()}} object immediately. When it is
resolved, it returns \code{TRUE} indicating consolidation success.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-vacuum"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-vacuum}{}}}
\subsection{Method \code{vacuum()}}{
Clean up consolidated fragments and array metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$vacuum(cfg = NULL, start_time = NULL,
  end_time = NULL, mode = c("fragments", "commits", "fragment_meta",
  "array_meta"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}} to set parameters
for the vacuum. When \code{NULL} (default) the configuration parameters
will be retrieve from object context.}

\item{\code{start_time, end_time}}{Optional time stamp values. A date time objects
of class \code{POSIXlt}. If not provided, the default values from configuration
object will be used.}

\item{\code{mode}}{The vacuum mode, one of the following:
\itemize{
\item \code{"fragments"}: - vacuum all fragments (default)
\item \code{"commits"}: - vacuum all commit files
\item \code{"fragment_meta"}: - vacuum only fragment metadata footers to a single file
\item \code{"array_meta"}: - vacuum array metadata only
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE}, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-vacuum_async"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-vacuum_async}{}}}
\subsection{Method \code{vacuum_async()}}{
Asynchronously clean up consolidated fragments and array metadata.

The clean up will run in a separate R process in a clean environment.

\strong{Note this function requires the \href{https://cran.r-project.org/web/packages/mirai/index.html}{mirai} package}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$vacuum_async(cfg = NULL, start_time = NULL,
  end_time = NULL, mode = c("fragments", "commits", "fragment_meta",
  "array_meta"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}} to set parameters
for the vacuum process. When \code{NULL} (default) the configuration parameters will
be retrieve from object context.}

\item{\code{start_time, end_time}}{Optional time stamp values. A date time objects
of class \code{POSIXlt}. If not provided, the default values from configuration
object will be used.}

\item{\code{mode}}{The vacuum mode, one of the following:
\itemize{
\item \code{"fragments"}: - vacuum all fragments (default)
\item \code{"commits"}: - vacuum all commit files
\item \code{"fragment_meta"}: - vacuum only fragment metadata footers to a single file
\item \code{"array_meta"}: - vacuum array metadata only
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This function will return a \code{\link[mirai:mirai]{mirai::mirai()}} object immediately. When it is
resolved, it returns \code{TRUE} indicating vacuum success.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-drop_attribute"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-drop_attribute}{}}}
\subsection{Method \code{drop_attribute()}}{
Remove an attribute from array.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$drop_attribute(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An attribute name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-frag_num"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-frag_num}{}}}
\subsection{Method \code{frag_num()}}{
The number of fragments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$frag_num()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-frag_to_vacuum"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-frag_to_vacuum}{}}}
\subsection{Method \code{frag_to_vacuum()}}{
Consolidated fragments to be removed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$frag_to_vacuum(trunc_uri = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trunc_uri}}{\code{TRUE} to truncate uri path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{data.frame} with four columns:
\itemize{
\item \code{Fragment}: the fragment index (starts at 1)
\item \code{start_timestamp}: fragment's start time stamp
\item \code{end_timestamp}: fragment's end time stamp
\item \code{URI}: fragment's truncated uri path (fragment name) when
\code{trunc_uri = TRUE} (default), otherwise the full uri path
}

Note that return object will be of class \code{data.table} if the
package is found in your system.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-frag_dump"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-frag_dump}{}}}
\subsection{Method \code{frag_dump()}}{
Dump to console the commit fragments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$frag_dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-frag_uris"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-frag_uris}{}}}
\subsection{Method \code{frag_uris()}}{
Return a \code{data.frame} with  time stamps and
fragments uris.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$frag_uris(trunc_uri = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trunc_uri}}{\code{TRUE} to truncate uri path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{data.frame} with four columns:
\itemize{
\item \code{Fragment}: the fragment index (start at 1)
\item \code{start_timestamp}: start time-stamp of when fragment was written
\item \code{end_timestamp}: end time-stamp of when fragment was written
\item \code{URI}: fragment's truncated uri path (fragment name) when
\code{trunc_uri = TRUE} (default), otherwise the full uri path
}

Note that return object will be of class \code{data.table} if the
package is found in your system.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-schema_upgrade"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-schema_upgrade}{}}}
\subsection{Method \code{schema_upgrade()}}{
Upgrade the array to the latest format version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$schema_upgrade(cfg = NULL, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A configuration object \code{\link[tiledb:tiledb_config]{tiledb::tiledb_config()}}.}

\item{\code{ctx}}{Optional \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}} object. By default, object's context
is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
