% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBURI.R
\name{TileDBURI}
\alias{TileDBURI}
\title{TileDB URI Parser}
\description{
An R6 class to parse and handle tileDB URI paths.
}
\examples{
uri <- "tiledb://namespace/s3://bucket"

uri_r6 <- TileDBURI$new(uri)

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{uri}}{The original URI}

\item{\code{scheme}}{The URI scheme}

\item{\code{tiledb_cloud_uri}}{URI for TileDB Cloud access}

\item{\code{object_uri}}{URI for direct access to the object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBURI-new}{\code{TileDBURI$new()}}
\item \href{#method-TileDBURI-is_remote_uri}{\code{TileDBURI$is_remote_uri()}}
\item \href{#method-TileDBURI-is_tiledb_cloud_uri}{\code{TileDBURI$is_tiledb_cloud_uri()}}
\item \href{#method-TileDBURI-is_tiledb_cloud_creation_uri}{\code{TileDBURI$is_tiledb_cloud_creation_uri()}}
\item \href{#method-TileDBURI-clone}{\code{TileDBURI$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBURI-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBURI-new}{}}}
\subsection{Method \code{new()}}{
Create a new TileDB URI.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBURI$new(uri)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI path for the TileDB object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBURI-is_remote_uri"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBURI-is_remote_uri}{}}}
\subsection{Method \code{is_remote_uri()}}{
Does the URI point to a remote object?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBURI$is_remote_uri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBURI-is_tiledb_cloud_uri"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBURI-is_tiledb_cloud_uri}{}}}
\subsection{Method \code{is_tiledb_cloud_uri()}}{
Is the URI prefix tiledb://?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBURI$is_tiledb_cloud_uri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBURI-is_tiledb_cloud_creation_uri"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBURI-is_tiledb_cloud_creation_uri}{}}}
\subsection{Method \code{is_tiledb_cloud_creation_uri()}}{
Is this a TileDB Cloud creation URI?
Example: "tiledb://namespace/s3://bucket"
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBURI$is_tiledb_cloud_creation_uri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBURI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBURI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBURI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
