% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfs_dir_tree.R
\name{vfs_dir_tree}
\alias{vfs_dir_tree}
\title{Print Directory Contents}
\usage{
vfs_dir_tree(uri, recursive = TRUE, vfs = NULL)
}
\arguments{
\item{uri}{URI path for the \code{TileDB} object.}

\item{recursive}{Should it recurse fully? Defaults to \code{TRUE}.}

\item{vfs}{A \code{\link[tiledb:tiledb_vfs]{tiledb::tiledb_vfs()}} object. If \code{NULL}
(default) will create a new VFS object.}
}
\value{
A character vector with file paths, invisibly.
}
\description{
A modified version of \code{\link[fs:dir_tree]{fs::dir_tree()}} to work with TileDB VFS.
}
\examples{
\dontrun{
# URI path
uri <- tempfile()

# Demo array
demo_array_UCBAdmissions(uri)

# Array instance
arrobj <- tdb_array(uri)

# Print directory contents
arrobj$dir_tree()
}
}
