% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfs_size.R
\name{vfs_size}
\alias{vfs_size}
\title{VFS Directory Size}
\usage{
vfs_size(uri, vfs = tiledb::tiledb_get_vfs())
}
\arguments{
\item{uri}{URI path for the \code{TileDB} object.}

\item{vfs}{A \code{\link[tiledb:tiledb_vfs]{tiledb::tiledb_vfs()}} object. Defaults to
\code{TileDB} VFS object from the \code{tiledb} environment and cache.}
}
\value{
The size of the directory.
}
\description{
VFS Directory Size
}
\examples{
\dontrun{
# URI path
uri <- tempfile()

# Demo array
demo_UCBAdmissions_array(uri)

# Directory size
vfs_size(uri)
}
}
