% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_timestamps.R
\name{array_timestamps}
\alias{array_timestamps}
\alias{array_timestamps.tiledb_array}
\alias{array_timestamps.TileDBArray}
\title{Get Array Timestamps}
\usage{
\method{array_timestamps}{tiledb_array}(object, tz = "", ...)

\method{array_timestamps}{TileDBArray}(object, tz = "", ...)
}
\arguments{
\item{object}{An \code{R} object that contains a \code{TileDB} resource pointer.}

\item{tz}{A character string for the time zone specification to be used
for the conversion in print method only. Defaults to \code{\link[=Sys.timezone]{Sys.timezone()}}.}

\item{...}{Other arguments passed to methods. Not used.}
}
\value{
An object of class \code{array_timestamps} that is a list that
holds the user supplied temporal timestamps and the time range the array
is opened at.
}
\description{
Get Array Timestamps
}
\seealso{
\code{\link[=group_timestamps]{group_timestamps()}}, \code{\link[=set_tiledb_timestamp]{set_tiledb_timestamp()}}
}
