% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{fetch_metadata}
\alias{fetch_metadata}
\alias{fetch_metadata.TileDBArray}
\alias{fetch_metadata.TileDBGroup}
\alias{fetch_metadata.tiledb_array}
\alias{fetch_metadata.tiledb_group}
\alias{fetch_metadata.character}
\title{Fetch TileDB Metadata}
\usage{
\method{fetch_metadata}{TileDBArray}(x, keys, timestamp = NULL)

\method{fetch_metadata}{TileDBGroup}(x, keys, timestamp = NULL)

\method{fetch_metadata}{tiledb_array}(x, keys, timestamp = NULL)

\method{fetch_metadata}{tiledb_group}(x, keys, timestamp = NULL)

\method{fetch_metadata}{character}(x, keys, timestamp = NULL)
}
\arguments{
\item{x}{An \code{R} object that points to a \code{TileDB} resource whose
metadata are to accessed.}

\item{keys}{A named list of key value metadata.}

\item{timestamp}{Optional datetime object of class \code{"POSIXct"} to write
at this timestamp.}
}
\value{
A named list of class \code{tdb_metadata}.
}
\description{
Fetch metadata as a list of key-value pairs for a \code{TileDB} array or group.
Optionally, you can access metadata at specific point in time
(time-travelling).
}
\details{
This functions work similar to \code{\link[=metadata]{metadata()}} but works with
a list of key value pairs and always returns a list.

The character method is intended for a valid URI path.

The methods will not alter the mode of the \code{TileDB} object; also, the object
will be opened temporarily to access the metadata if it is closed.
}
\seealso{
\code{\link[=set_metadata]{set_metadata()}} and \code{\link[=metadata]{metadata()}}.
}
