% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_write.R
\name{open_write}
\alias{open_write}
\alias{open_write.TileDBArray}
\alias{open_write.tiledb_array}
\alias{open_write.TileDBGroup}
\alias{open_write.tiledb_group}
\alias{open_write.character}
\title{Open a TileDB Resource at Write Mode}
\usage{
\method{open_write}{TileDBArray}(object, timestamp = NULL, ...)

\method{open_write}{tiledb_array}(object, timestamp = NULL, ...)

\method{open_write}{TileDBGroup}(object, timestamp = NULL, ...)

\method{open_write}{tiledb_group}(object, timestamp = NULL, ...)

\method{open_write}{character}(object, timestamp = NULL, ...)
}
\arguments{
\item{object}{An \code{R} object that contains a \code{TileDB} resource pointer.}

\item{timestamp}{Optional datetime object of class \code{"POSIXct"} to write
at this timestamp.}

\item{...}{Other arguments passed to methods. Not used.}
}
\value{
An object of class \code{tiledb_array} or \code{tiledb_group} depending on
the method.
}
\description{
Open a URI, \code{\link[tiledb:tiledb_array]{tiledb::tiledb_array()}}, \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group()}}, \link{TileDBArray}
or \link{TileDBGroup} at write mode with an optional timestamp.
}
\details{
Opening a character string should be a valid URI path for a TileDB resource.

Objects other than a URI character are implicitly closed if found opened and
then opened again at write mode.
}
