% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6.tiledb-package.R
\docType{package}
\name{R6.tiledb-package}
\alias{R6.tiledb}
\alias{R6.tiledb-package}
\title{R6.tiledb: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cgiachalis/R6.tiledb}
  \item Report bugs at \url{https://github.com/cgiachalis/R6.tiledb/issues}
}

}
\author{
\strong{Maintainer}: Constantinos Giachalis \email{cgiachalis@outlook.com}

}
\keyword{internal}
