% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tiledb_timestamp.R
\name{set_tiledb_timestamp}
\alias{set_tiledb_timestamp}
\title{Set a \code{TileDB} timestamp}
\usage{
set_tiledb_timestamp(start_time, end_time, tz = "")
}
\arguments{
\item{start_time, end_time}{An object coercible to \code{POSIXct}. See \code{\link[=as.POSIXct]{as.POSIXct()}}.}

\item{tz}{A character string for the time zone specification to be used
for the conversion. Defaults to \code{\link[=Sys.timezone]{Sys.timezone()}}.}
}
\value{
An list object of class \code{tiledb_timestamp}.
}
\description{
Define a list of start,end timestamps intended for opening a \code{TileDB}
object, either at initialisation or via active field \code{tiledb_timestamp}.
}
\examples{
# Character input
set_tiledb_timestamp("2025-01-01", "2025-08-01", tz = "UTC")

# Numeric input
set_tiledb_timestamp(1000000, 1000000*4, tz = "UTC")

# Date end point
set_tiledb_timestamp(end_time = as.Date("1990-01-01"), tz = "UTC")

# Default
set_tiledb_timestamp(tz = "Europe/London")

# Invalid: start_time > end_time
# set_tiledb_timestamp(start_time = 1, end_time = 0)
# Error: `start_time` is greater than `end_time`.

}
\seealso{
\link{TileDBObject}
}
