% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBArrayExtended.R
\name{TileDBArrayExtended}
\alias{TileDBArrayExtended}
\title{TileDB Array Extended Class}
\value{
An object \code{TileDBArrayExtented} of class \code{R6}.
}
\description{
An R6 class for representing an extended TileDB Array.
\subsection{Initialization}{

A new \code{TileDBArrayExtended} instance is initialized using the \code{new()} method:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ # uri path
 uri <- tempdir()
 # new instance
 obj <- TileDBArrayExtended$new(uri = uri)
 # does array exist at this uri
 obj$exists() # FALSE

 unlink(uri)
}\if{html}{\out{</div>}}
}
}
\section{Super classes}{
\code{\link[R6.tiledb:TileDBObject]{R6.tiledb::TileDBObject}} -> \code{\link[R6.tiledb:TileDBArray]{R6.tiledb::TileDBArray}} -> \code{TileDBArrayExtended}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fragments_object}}{Access an \link{TileDBFragments} instance.}

\item{\code{schema_version}}{Retrieve the schema version for this array.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBArrayExtended-new}{\code{TileDBArrayExtended$new()}}
\item \href{#method-TileDBArrayExtended-enum_columns}{\code{TileDBArrayExtended$enum_columns()}}
\item \href{#method-TileDBArrayExtended-enum_levels}{\code{TileDBArrayExtended$enum_levels()}}
\item \href{#method-TileDBArrayExtended-has_enums}{\code{TileDBArrayExtended$has_enums()}}
\item \href{#method-TileDBArrayExtended-consolidate}{\code{TileDBArrayExtended$consolidate()}}
\item \href{#method-TileDBArrayExtended-vacuum}{\code{TileDBArrayExtended$vacuum()}}
\item \href{#method-TileDBArrayExtended-finfo_to_vacuum}{\code{TileDBArrayExtended$finfo_to_vacuum()}}
\item \href{#method-TileDBArrayExtended-finfo_dump}{\code{TileDBArrayExtended$finfo_dump()}}
\item \href{#method-TileDBArrayExtended-finfo_uris}{\code{TileDBArrayExtended$finfo_uris()}}
\item \href{#method-TileDBArrayExtended-schema_upgrade}{\code{TileDBArrayExtended$schema_upgrade()}}
\item \href{#method-TileDBArrayExtended-clone}{\code{TileDBArrayExtended$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="class"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-class'><code>R6.tiledb::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="exists"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-exists'><code>R6.tiledb::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="is_open"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-is_open'><code>R6.tiledb::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="mode"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-mode'><code>R6.tiledb::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="reopen"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-reopen'><code>R6.tiledb::TileDBObject$reopen()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="attributes"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-attributes'><code>R6.tiledb::TileDBArray$attributes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="attrnames"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-attrnames'><code>R6.tiledb::TileDBArray$attrnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="close"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-close'><code>R6.tiledb::TileDBArray$close()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="colnames"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-colnames'><code>R6.tiledb::TileDBArray$colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="dimensions"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-dimensions'><code>R6.tiledb::TileDBArray$dimensions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="dimnames"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-dimnames'><code>R6.tiledb::TileDBArray$dimnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="get_metadata"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-get_metadata'><code>R6.tiledb::TileDBArray$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="open"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-open'><code>R6.tiledb::TileDBArray$open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="print"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-print'><code>R6.tiledb::TileDBArray$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="schema"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-schema'><code>R6.tiledb::TileDBArray$schema()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="schema_info"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-schema_info'><code>R6.tiledb::TileDBArray$schema_info()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="set_metadata"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-set_metadata'><code>R6.tiledb::TileDBArray$set_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBArray" data-id="tiledb_array"><a href='../../R6.tiledb/html/TileDBArray.html#method-TileDBArray-tiledb_array'><code>R6.tiledb::TileDBArray$tiledb_array()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TileDBArrayExtended} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$new(uri, ctx = NULL, tiledb_timestamp = NULL,
  internal_use = getOption("R6.tiledb.internal"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI path for the TileDB Array.}

\item{\code{ctx}}{Optional \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}} object.}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp.}

\item{\code{internal_use}}{A character value that gives access to new instance.
Use \code{options(R6.tiledb.internal = NULL)} for internal mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-enum_columns"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-enum_columns}{}}}
\subsection{Method \code{enum_columns()}}{
Retrieve factor columns (attributes).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$enum_columns()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector with factor columns (enum attributes).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-enum_levels"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-enum_levels}{}}}
\subsection{Method \code{enum_levels()}}{
Retrieve factor levels for a given attribute.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$enum_levels(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{An attribute name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector with levels (enum values).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-has_enums"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-has_enums}{}}}
\subsection{Method \code{has_enums()}}{
Checks if array has factor columns.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$has_enums()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A boolean. \code{TRUE} indicating the array has factor columns and
\code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-consolidate"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-consolidate}{}}}
\subsection{Method \code{consolidate()}}{
Fragment consolidation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$consolidate(cfg = NULL, start_time, end_time,
  ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A \code{TileDB} configuration object.}

\item{\code{start_time, end_time}}{Optional time stamp value. A date time object
of class \code{POSIXlt}. If not provided,the default values from configuration
object will be used.}

\item{\code{ctx}}{A \code{TileDB} context object. By default, object's context is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-vacuum"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-vacuum}{}}}
\subsection{Method \code{vacuum()}}{
Remove consolidated fragments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$vacuum(cfg = NULL, start_time, end_time, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A \code{TileDB} configuration object.}

\item{\code{start_time, end_time}}{Optional time stamp value. A date time object
of class \code{POSIXlt}. If not provided,the default values from configuration
object will be used.}

\item{\code{ctx}}{A \code{TileDB} context object. By default, object's context is used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-finfo_to_vacuum"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-finfo_to_vacuum}{}}}
\subsection{Method \code{finfo_to_vacuum()}}{
Consolidated fragments to be removed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$finfo_to_vacuum(trunc_uri = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trunc_uri}}{\code{TRUE} to truncate uri path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-finfo_dump"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-finfo_dump}{}}}
\subsection{Method \code{finfo_dump()}}{
Dump to console the commit fragments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$finfo_dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-finfo_uris"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-finfo_uris}{}}}
\subsection{Method \code{finfo_uris()}}{
Fragments uri and time stamps.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$finfo_uris(trunc_uri = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{trunc_uri}}{\code{TRUE} to truncate uri path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-schema_upgrade"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-schema_upgrade}{}}}
\subsection{Method \code{schema_upgrade()}}{
Upgrade array schema
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$schema_upgrade(cfg = NULL, ctx = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cfg}}{A \code{TileDB} configuration object.}

\item{\code{ctx}}{A \code{TileDB} context object. By default, object's context is used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBArrayExtended-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBArrayExtended-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBArrayExtended$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
