% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBGroup.R
\name{TileDBGroup}
\alias{TileDBGroup}
\title{TileDB Group Base Class}
\value{
An object \code{TileDBGroup} of class \code{R6}.
}
\description{
Base class for interacting with TileDB Groups.
\subsection{Initialization}{

A new \code{TileDBGroup} instance is initialized using the \code{new()} method:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{ # uri path
 uri <- tempdir()
 # new instance
 obj <- TileDBGroup$new(uri = uri)
 # does array exist at this uri
 obj$exists() # FALSE

 unlink(uri)
}\if{html}{\out{</div>}}
}
}
\section{Super class}{
\code{\link[R6.tiledb:TileDBObject]{R6.tiledb::TileDBObject}} -> \code{TileDBGroup}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{object}}{Access the underlying \code{\link[tiledb:tiledb_group]{tiledb::tiledb_group()}} object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBGroup-new}{\code{TileDBGroup$new()}}
\item \href{#method-TileDBGroup-create}{\code{TileDBGroup$create()}}
\item \href{#method-TileDBGroup-open}{\code{TileDBGroup$open()}}
\item \href{#method-TileDBGroup-close}{\code{TileDBGroup$close()}}
\item \href{#method-TileDBGroup-remove}{\code{TileDBGroup$remove()}}
\item \href{#method-TileDBGroup-count_members}{\code{TileDBGroup$count_members()}}
\item \href{#method-TileDBGroup-get_members_df}{\code{TileDBGroup$get_members_df()}}
\item \href{#method-TileDBGroup-get_member}{\code{TileDBGroup$get_member()}}
\item \href{#method-TileDBGroup-set_member}{\code{TileDBGroup$set_member()}}
\item \href{#method-TileDBGroup-get_metadata}{\code{TileDBGroup$get_metadata()}}
\item \href{#method-TileDBGroup-set_metadata}{\code{TileDBGroup$set_metadata()}}
\item \href{#method-TileDBGroup-names}{\code{TileDBGroup$names()}}
\item \href{#method-TileDBGroup-to_list}{\code{TileDBGroup$to_list()}}
\item \href{#method-TileDBGroup-member_exists}{\code{TileDBGroup$member_exists()}}
\item \href{#method-TileDBGroup-print}{\code{TileDBGroup$print()}}
\item \href{#method-TileDBGroup-dump}{\code{TileDBGroup$dump()}}
\item \href{#method-TileDBGroup-clone}{\code{TileDBGroup$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="class"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-class'><code>R6.tiledb::TileDBObject$class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="exists"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-exists'><code>R6.tiledb::TileDBObject$exists()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="is_open"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-is_open'><code>R6.tiledb::TileDBObject$is_open()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="mode"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-mode'><code>R6.tiledb::TileDBObject$mode()</code></a></span></li>
<li><span class="pkg-link" data-pkg="R6.tiledb" data-topic="TileDBObject" data-id="reopen"><a href='../../R6.tiledb/html/TileDBObject.html#method-TileDBObject-reopen'><code>R6.tiledb::TileDBObject$reopen()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TileDBGroup} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$new(uri, ctx = NULL, tiledb_timestamp = NULL,
  internal_use = getOption("R6.tiledb.internal"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI path for the TileDB Group.}

\item{\code{ctx}}{Optional \code{\link[tiledb:tiledb_ctx]{tiledb::tiledb_ctx()}} object.}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp.}

\item{\code{internal_use}}{A character value that gives access to new instance.
Use \code{options(R6.tiledb.internal = NULL)} for internal mode.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-create"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-create}{}}}
\subsection{Method \code{create()}}{
Create a TileDB Group object given the class URI path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$create(mode = "WRITE")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{Mode to open : either \code{"READ"} or \code{"WRITE"} (default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-open"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-open}{}}}
\subsection{Method \code{open()}}{
Open TileDB group object for read or write.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$open(mode = c("READ", "WRITE"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{Mode to open : either \code{"READ"} or \code{"WRITE"}.  Default is \code{"READ"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-close"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-close}{}}}
\subsection{Method \code{close()}}{
Close the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$close()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-remove"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-remove}{}}}
\subsection{Method \code{remove()}}{
Remove member.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$remove(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the member to remove.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} value, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-count_members"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-count_members}{}}}
\subsection{Method \code{count_members()}}{
Count the number of members in the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$count_members()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of members in the group.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_members_df"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_members_df}{}}}
\subsection{Method \code{get_members_df()}}{
List the members of the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_members_df(type = c("ALL", "GROUP", "ARRAY"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{Select type member, either\code{"ALL"}, \code{"GROUP"}
or \code{"ARRAY"}. By default all member types are listed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{data.frame} with columns \code{uri}, \code{type}, and \code{name}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_member"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_member}{}}}
\subsection{Method \code{get_member()}}{
Retrieve a group member by name. If the member isn't already
open, it is opened in the same mode as the parent.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_member(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the member.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{TileDBArray} or \code{TileDBGroup}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-set_member"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-set_member}{}}}
\subsection{Method \code{set_member()}}{
Add new member to the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$set_member(object, name = NULL, relative = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object}}{A \code{TileDBArray} or \code{TileDBGroup} object to add.}

\item{\code{name}}{Name to use for the member. By default the base name of
the object's URI is used.}

\item{\code{relative}}{An optional logical value indicating whether the new
object's URI is relative to the group's URI. If \code{NULL} (the
default), the object's URI is assumed to be relative unless it is a
\verb{tiledb://} URI.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} value, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Retrieve metadata from the TileDB group.

The \code{TileDBGroup} will be opened for \code{"READ"} to fetch its metadata.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$get_metadata(key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The name of the metadata attribute to retrieve.
The default \code{NULL} returns all metadata.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The key metadata value or a \code{list} of all metadata
values when \code{NULL}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-set_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-set_metadata}{}}}
\subsection{Method \code{set_metadata()}}{
Add list of metadata.

\code{TileDBGroup} should be open for \code{"WRITE"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$set_metadata(metadata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metadata}}{Named list of metadata to add.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL} value, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-names"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-names}{}}}
\subsection{Method \code{names()}}{
Retrieve the members' names.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$names()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{character} vector of member names.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-to_list}{}}}
\subsection{Method \code{to_list()}}{
Retrieve a \code{list} of members.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{list}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-member_exists"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-member_exists}{}}}
\subsection{Method \code{member_exists()}}{
Check if a member exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$member_exists(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of the member to check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A logical value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-print"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-print}{}}}
\subsection{Method \code{print()}}{
Print summary of the group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-dump"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-dump}{}}}
\subsection{Method \code{dump()}}{
Dump the TileDB Group structure to string.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$dump(recursive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recursive}}{Should the nested uris be returned recursively?
Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{character} string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBGroup-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBGroup-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBGroup$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
