% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBObject.R
\name{TileDBObject}
\alias{TileDBObject}
\title{TileDB Object Base Class}
\description{
Base class to implement shared functionality across the \code{TileDBArray} and
\code{TileDBGroup} classes.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ctx}}{TileDB Context.}

\item{\code{tiledb_timestamp}}{Time that object was opened at.}

\item{\code{uri}}{The URI of the TileDB object.}

\item{\code{object_type}}{The TileDB object type.

TileDB object types:
\itemize{
\item \code{"ARRAY"}, for dense or sparse array resource
\item \code{"GROUP"}, for group resource
\item \code{"INVALID"}, for not a TileDB resource
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TileDBObject-new}{\code{TileDBObject$new()}}
\item \href{#method-TileDBObject-class}{\code{TileDBObject$class()}}
\item \href{#method-TileDBObject-is_open}{\code{TileDBObject$is_open()}}
\item \href{#method-TileDBObject-mode}{\code{TileDBObject$mode()}}
\item \href{#method-TileDBObject-reopen}{\code{TileDBObject$reopen()}}
\item \href{#method-TileDBObject-exists}{\code{TileDBObject$exists()}}
\item \href{#method-TileDBObject-clone}{\code{TileDBObject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-new"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-new}{}}}
\subsection{Method \code{new()}}{
Create a new TileDB object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$new(uri, ctx = NULL, tiledb_timestamp = NULL,
  internal_use = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{URI path for the TileDB object.}

\item{\code{ctx}}{Optional \code{\link[=tiledb_ctx]{tiledb_ctx()}} object.}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp.}

\item{\code{internal_use}}{A character value that gives access to internal method \code{new()}.
The \code{new()} method should not be called directly.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-class"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-class}{}}}
\subsection{Method \code{class()}}{
Print the name of the R6 class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$class()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-is_open"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-is_open}{}}}
\subsection{Method \code{is_open()}}{
Determine if the object is open for reading or writing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$is_open()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the object is open, otherwise \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-mode"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-mode}{}}}
\subsection{Method \code{mode()}}{
Get the mode of the object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$mode()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
If the object is closed, returns \code{"CLOSED"};
otherwise returns the mode (e.g. \code{"READ"}) of the object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-reopen"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-reopen}{}}}
\subsection{Method \code{reopen()}}{
Close and reopen the TileDB object in a new mode.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$reopen(mode = c("READ", "WRITE"), tiledb_timestamp = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mode}}{New mode to open the object in; choose from: \code{"READ"} or \code{"WRITE"}.}

\item{\code{tiledb_timestamp}}{Optional Datetime (POSIXct) with TileDB timestamp.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisibly returns \code{self} opened in \code{mode}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-exists"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-exists}{}}}
\subsection{Method \code{exists()}}{
Check if the object exists.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$exists()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the object exists, \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TileDBObject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TileDBObject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TileDBObject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
