% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vfs_dir_tree.R
\name{vfs_dir_tree}
\alias{vfs_dir_tree}
\title{Print directory contents}
\usage{
vfs_dir_tree(uri, vfs = tiledb::tiledb_get_vfs())
}
\arguments{
\item{uri}{URI path for the \code{TileDB} object.}

\item{vfs}{A \code{\link[tiledb:tiledb_vfs]{tiledb::tiledb_vfs()}} object. Defaults to
\code{TileDB} VFS object from the \code{tiledb} environment and cache.}
}
\value{
A character vector with file paths, invisibly.
}
\description{
Print directory contents
}
\examples{
\dontrun{
# URI path
uri <- tempfile()

# Demo array
demo_UCBAdmissions_array(uri)

# Array instance
arrobj <- tdb_array(uri)

# Print directory contents
arrobj$dir_tree()
}
}
